Several changes needed to make the first version of the aprc tool work.
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/SCHEDULING@127633 523d945c-050c-4681-91ec-863ad3bb968a

+++ Common/src/alma/scheduling/datamodel/output/dao/XmlOutputDaoImpl.java
@@ -66,7 +66,7 @@ public class XmlOutputDaoImpl implements OutputDao {
         for(int i = 0; i < op.length; i++){
             ObservationProject tmpOp = itOp.next();
             op[i].setExecutionTime(tmpOp.getExecutionTime()); 
-            op[i].setScienceRating(tmpOp.getScienceRating());
+            op[i].setScienceRating((int) tmpOp.getScienceRating());
             op[i].setStatus(alma.scheduling.output.generated.types.ExecutionStatus.valueOf(
                     tmpOp.getStatus().name()));
             //set the affiliations
